fix: address CodeRabbit review comments on CIVITAS integration
- CivitasDataMapper: omit external_id when missing, recursive hash normalization, null for missing metric values
- NgsiLdClient: validate OAuth token before returning
- SensorThingsClient: cast total to int
- SyncService: query all 4 models for last sync status, throw in stub syncAll
- Remove sync provenance fields from $fillable on all models
- SyncCivitasData: add try/catch per tenant, track exit code
- ManageIntegrations: enforce config check in canAccess()
- setup.sh: conditional FROST_DB_PASSWORD, remove || true, add Java version check
- README: add git/curl to prerequisites
- Migration to make metric_values.value nullable

+++ app/Services/Integration/SensorThingsClient.php
@@ -84,7 +84,7 @@ class SensorThingsClient implements ExternalDataSourceInterface
 
         return [
             'entities' => $data['value'] ?? [],
-            'total' => $data['@iot.count'] ?? count($data['value'] ?? []),
+            'total' => (int) ($data['@iot.count'] ?? count($data['value'] ?? [])),
         ];
     }
